+++ spec/integration/resource/api500/c7000/logical_enclosure/delete_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+klass = OneviewSDK::API500::C7000::LogicalEnclosure
+RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
+  let(:current_client) { $client_500 }
+  include_examples 'LogicalEnclosureDeleteExample', 'integration api500 context', 'C7000'
+end